relocation of some GUI classes to separate files
