+++ pages/main.json
@@ -18,7 +18,7 @@
             "background": [0, 0, 0],
             "actions": {
                 "on-press": ["Media:pauseplay"],
-                "on-release": ["Media:pauseplay"]
+                "on-release": []
             }
         },
         "1x1": {
